requirements.txt
django==1.9.5
tweepy==3.5.0
psycopg2==2.6.1
pytz
djangorestframework==3.3.3